finished, 能是实现所有的PID调参实验，IMU解算单独更新，多种模式并存

+++ HARDWARE/BaseTimer/base_timer_wrapper.c
@@ -16,9 +16,9 @@ void delay_us(u32 nus)
 	base_timer_delay_us(nus);
 }
 
-uint32_t micros()
+uint64_t micros()
 {
-	uint32_t micros = base_timer_get_time() * 1000 + base_timer_get_us();
+	uint64_t micros = base_timer_get_time() * 1000 + base_timer_get_us();
 	return micros;
 }
 //end of file
